requirements.txt
# Python version requirement
# Python 3.13.3 

# API framework
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.46.2

# LLM providers
openai==1.93.0
anthropic==0.56.0
google-generativeai==0.8.5

# Data processing
pandas==2.3.0
openpyxl==3.1.5
numpy==2.2.6

# Database
snowflake-connector-python==3.15.0

# SQL parsing and analysis
sqlglot==24.0.0

# Configuration and environment
python-dotenv==1.1.0

# Utilities
requests==2.32.3
json5==0.9.0
pytz==2025.2

# HTTP and API utilities
httpx==0.28.1
httpcore==1.0.9
anyio==4.9.0

# Chart generation support
matplotlib==3.10.3
seaborn==0.13.2

# Authentication and security
cryptography==45.0.3
pyOpenSSL==25.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
